helios: Minor fix to GOES flare classification

+++ helios/src/main/scala-2.11/io/github/mandar2812/PlasmaML/helios/data/GOESData.scala
@@ -55,7 +55,7 @@ object GOESData {
     * 3 - X
     *
     * */
-  def getFlareClass(flux_1_8_ang: Double): Int =
+  def getFlareClass(flux_1_8_ang: Float): Int =
     if(flux_1_8_ang < -6d) 0
     else if(flux_1_8_ang >= -6d && flux_1_8_ang < -5d) 1
     else if(flux_1_8_ang >= -5d && flux_1_8_ang <= -4d) 2
